Add files via upload
Put this in a separate branch in case using packages and imports messes everyone up. Put the Constants and builtFloorPlans in a folder called resources. This is the name of the package and it is being imported by FloorPlans, FinderApp, and Map. 
The 6th floor in TFDL is still a work in progress.